requirements.txt
fast-mcp
langchain-huggingface>=0.1.0
langchain-community>=0.3.0
faiss-cpu>=1.7.4
pillow>=10.0.0
sentence-transformers>=2.2.0
